refactor(grading): 拆分评委契约并移除遗留结果类型
