+++ src/grading/contracts/index.ts
@@ -1,0 +1,2 @@
+export type * from './config.js';
+export type * from './result.js';
