+++ src/grading/judge.ts
@@ -1,6 +1,11 @@
+import type {
+  DimensionResult,
+  EnsembleJudgeResult,
+  JudgeAgreement,
+} from './contracts/result.js';
+import type { JudgeConfig } from './contracts/config.js';
 import type { ExecutorFn } from '../executors/contracts/ports.js';
 import type { ToolCallInfo, TurnInfo } from '../executors/contracts/trace.js';
-import type { DimensionResult, EnsembleJudgeResult, JudgeAgreement, JudgeConfig } from '../types/index.js';
 import { buildJudgePrompt, JUDGE_SYSTEM_PROMPT } from '../shared/llm-prompts/judge-prompts.js';
 import { buildJudgeTraceSummary } from '../shared/llm-prompts/judge-trace.js';
 // 评分类 prompt 已收口到 shared/llm-prompts/judge-prompts.ts(单一来源 + prompt-registry 冻结)。
