+++ src/types/index.ts
@@ -1,5 +1,4 @@
 export * from './shared.js';
-export * from './judge.js';
 export * from './doctor.js';
 export * from './diagnosis.js';
 export * from './skill-index.js';
