+++ test/scripts/legacy-evaluation-removal.test.ts
@@ -7,6 +7,7 @@ const FORBIDDEN_FILES = [
   'src/types/report.ts',
   'src/types/storage.ts',
   'src/types/eval.ts',
+  'src/types/judge.ts',
   'src/eval-workflows/run-evaluation.ts',
   'src/eval-workflows/evaluation-pipeline.ts',
   'src/eval-workflows/batch-evaluation-workflow.ts',
@@ -23,6 +24,7 @@ const FORBIDDEN_SPECIFIER_FRAGMENTS = [
   '/types/report',
   '/types/storage',
   '/types/eval',
+  '/types/judge',
   '/eval-workflows/run-evaluation',
   '/eval-workflows/evaluation-pipeline',
   '/server/report-store',
